PHP8対応でDocker未対応及びインストーラーが動作していない #908  (#909)
* DockerのコンテナのPHPを8.3に変更
* 想定しないファイルの上書きが発生していたため修正

+++ setup/scripts/10_Delete_Modules.php
@@ -23,15 +23,21 @@ require_once 'includes/Loader.php';
 $db = PearDatabase::getInstance();
 
 $module = Vtiger_Module::getInstance('ExtensionStore');
-$module->delete();
+if($module) {
+    $module->delete();
+}
 
 $db->query("delete from vtiger_links where linklabel like 'ExtensionStoreCommonHeaderScript'");
 
 $module = Vtiger_Module::getInstance('PBXManager');
-$module->delete();
+if($module) {
+    $module->delete();
+}
 
 $module = Vtiger_Module::getInstance('Google');
-$module->delete();
+if($module) {
+    $module->delete();
+}
 
 $db->query("delete from vtiger_settings_blocks where label like 'LBL_EXTENSIONS'");
